scheduler.py
from apscheduler.schedulers.background import BackgroundScheduler
from datetime import datetime
import json
import requests
from dotenv import load_dotenv
from config import TOKEN, CHAT_ID
load_dotenv()

def send_scheduled_message():
    now = datetime.now().strftime("%Y-%m-%d %H:%M")
    print(f"🕒 Checking for scheduled messages at {now}")
    try:
        with open("data/messages_by_date.json", "r", encoding='utf-8') as f:
            messages = json.load(f)
        if now in messages:
            print(f"Sending ID:{CHAT_ID}")
            text = messages[now]
            print(f"📨 Sending: {text}")
            url = f"https://api.telegram.org/bot{TOKEN}/sendMessage"
            print(f"URL:{url}")
            data = {"chat_id": CHAT_ID, "text": text}
            response = requests.post(url, data=data)
            print(response.status_code)
            print(response.text)
            return response
    except Exception as e:
        print(f"⚠️ Error in scheduler: {e}")

def start_scheduler():
    scheduler = BackgroundScheduler()
    scheduler.add_job(send_scheduled_message, 'interval', minutes=1)
    scheduler.start()
