requirements.txt
python-telegram-bot==20.7
python-dotenv
Flask==2.3.3
requests==2.32.4
APScheduler==3.10.4
pytest